cookbook/pocketflow-parallel-batch-flow/requirements.txt
pocketflow
Pillow>=10.0.0  # For image processing
numpy>=1.24.0   # For image array operations 